[feat] #20 feat : 기록보기 메인 완료

+++ data/src/main/java/com/masshow/data/model/response/MonthlyRecordResponse.kt
@@ -11,7 +11,7 @@ data class MonthlyRecordResponse(
 
 data class History(
     val drankAt: String,
-    val historyId: Int,
+    val historyId: Long,
     val liquorDetailNames: List<String>
 )
